[JAXRS-CXF] Add regular constructors to POJOs
Before this change, POJOs were generated without regular constructors.
This would mean you have to use the default constructor and setters or
builder-style methods. Some people consider regular constructors safer
to use due to compiliation failing when changing members of an object.
This change adds a regular constructor to all POJOs, and specifically to
the generated models.

Three constructors are now generated:
(1) An empty constructor (needed for backward compatiblity and for Jackson)
(2) A constructor with only required parameters as arguments
(3) A constructor for both required and optional parameters